try renew GPA

+++ GPACalculatorAPI/Models/Requests/CreateSubjectRequest.cs
@@ -4,6 +4,6 @@ namespace GPACalculatorAPI.Models.Requests
     public class CreateSubjectRequest
     {
         public string Name { get; set; }
-        public double Credit { get; set; }
+        public int Credit { get; set; }
     }
 }
